+++ src/presentation/controllers/login/login.spec.ts
@@ -4,7 +4,7 @@ import {
   ok,
   serverError,
   unauthorized,
-} from '../../helpers/http-helper'
+} from '../../helpers/http/http-helper'
 import { HttpRequest, Authentication } from './login-protocols'
 import { LoginController } from './login'
 import { Validation } from '../signup/signup-protocols'
