+++ src/presentation/controllers/login/login.ts
@@ -4,7 +4,7 @@ import {
   ok,
   serverError,
   unauthorized,
-} from '../../helpers/http-helper'
+} from '../../helpers/http/http-helper'
 import { Validation } from '../signup/signup-protocols'
 import {
   Authentication,
