+++ src/presentation/controllers/signup/signup.ts
@@ -1,4 +1,4 @@
-import { badRequest, ok, serverError } from '../../helpers/http-helper'
+import { badRequest, ok, serverError } from '../../helpers/http/http-helper'
 import {
   AddAccount,
   Controller,
